requirements.txt
httplib2==0.9.1
oauth2==1.5.211
py-trello==0.1.6
